abstract out createSlot and createSlots
